rename liftover util file + move 'parse_and_rebuild_response' into a general util file
