+++ src/anyvar/utils/general.py
@@ -1,0 +1,31 @@
+"""General utility functions, etc."""
+
+import json
+from typing import cast
+
+from fastapi import Response
+from fastapi.responses import JSONResponse, StreamingResponse
+
+
+async def parse_and_rebuild_response(
+    response: StreamingResponse,
+) -> tuple[dict, Response]:
+    """Convert a `Response` object to a dict, then re-build a new Response object (since parsing exhausts the Response `body_iterator`).
+
+    :param response: the `Response` object to parse
+    :return: a tuple with a dictionary representation of the Response and a new `Response` object
+    """
+    response_chunks: list[bytes] = [
+        cast(bytes, chunk) async for chunk in response.body_iterator
+    ]
+    response_body_encoded = b"".join(response_chunks)
+    response_body = response_body_encoded.decode("utf-8")
+    response_json = json.loads(response_body)
+
+    new_response = JSONResponse(
+        content=response_json,
+        status_code=response.status_code,
+        media_type=response.media_type,
+    )
+
+    return (response_json, new_response)
